Put IEEE802154 in separate iface module
